Moved the user-related components to a new User Module.
