+++ src/app/user/components/iconEditor/iconEditor.component.ts
