+++ src/app/user/components/messages/messages.component.spec.ts
@@ -46,7 +46,7 @@ import { By } from "@angular/platform-browser";
 import { NO_ERRORS_SCHEMA } from "@angular/core";
 
 import { AppMessaging } from "./messages.component";
-import { AuthService } from "../../common/services/auth.service";
+import { AuthService } from "../../../common/services/auth.service";
 import { mockAuthedUser } from "@tests/mockData";
 import { FullThread } from "@app/interfaces/thread.interface";
 import { MockDeleteForm } from "@tests/mockForms";
